feat(address) serialize/deserialize using bech32
